Added a 12x slider
I tested with "* 4" and it seemed a tad small for the Config tweak mod (using Patch Manager to add kos to cockpits)

So Working in the Script i noticed a block of math keeping it Clamped. so i left it as 8*BaseDiskSpace and increased smoothing

cost is not an issue without a tweak patch, so that has not been changed but using 1024 * 8 with extra cost would be recommended to follow IRL ByteSizes. Most parts in the KOS mod seem to rise just out of that area. i felt as if it was to limited for upgrades

One last note: would recommend adjusting "public float ECPerInstruction = 0.000004F;" to 0.000008F.